chore: preliminary clean up and reorg of cycle_group (#16645)
Preliminary `cycle_group` cleanup (no logic changes):
- Extract subclasses `cycle_scalar`, `straus_lookup_table` and
`straus_scalar_slice` and move to their own files
- Remove methods marked `requires IsNotUltraArithmetic<Builder>` since
no no-Ultra builders exist

Note: some branching based on "is Ultra" remains _within_ certain
methods and will be removed in subsequent PRs
