+++ barretenberg/cpp/src/barretenberg/stdlib/primitives/group/cycle_group.cpp
@@ -24,6 +24,7 @@ namespace bb::stdlib {
  * @note Please don't use this constructor in case you want to assign the
  * coordinates later.
  */
+// WORKTODO: is this fuzzer only logic? if so lets mark it accordingly e.g. with sig (Builder* _context, FUZZER_ONLY)
 template <typename Builder>
 cycle_group<Builder>::cycle_group(Builder* _context)
     : x(0)
@@ -338,24 +339,6 @@ template <typename Builder> void cycle_group<Builder>::standardize()
 }
 
 /**
- * @brief Evaluates a doubling. Does not use Ultra double gate
- *
- * @tparam Builder
- * @param unused param is due to interface-compatibility with the UltraArithmetic version of `dbl`
- * @return cycle_group<Builder>
- */
-template <typename Builder>
-cycle_group<Builder> cycle_group<Builder>::dbl([[maybe_unused]] const std::optional<AffineElement> /*unused*/) const
-    requires IsNotUltraArithmetic<Builder>
-{
-    auto modified_y = field_t::conditional_assign(is_point_at_infinity(), 1, y);
-    auto lambda = (x * x * 3) / (modified_y + modified_y);
-    auto x3 = lambda.madd(lambda, -x - x);
-    auto y3 = lambda.madd(x - x3, -modified_y);
-    return cycle_group(x3, y3, is_point_at_infinity());
-}
-
-/**
  * @brief Evaluates a doubling. Uses Ultra double gate
  *
  * @tparam Builder
@@ -424,33 +407,6 @@ cycle_group<Builder> cycle_group<Builder>::dbl(const std::optional<AffineElement
     // We need to manually propagate the origin tag
     result.x.set_origin_tag(OriginTag(x.get_origin_tag(), y.get_origin_tag()));
     result.y.set_origin_tag(OriginTag(x.get_origin_tag(), y.get_origin_tag()));
-    return result;
-}
-
-/**
- * @brief Will evaluate ECC point addition over `*this` and `other`.
- *        Incomplete addition formula edge cases are *NOT* checked!
- *        Only use this method if you know the x-coordinates of the operands cannot collide
- *        and none of the operands is a point at infinity
- *        Standard version that does not use ecc group gate
- *
- * @tparam Builder
- * @param other
- * @return cycle_group<Builder>
- */
-template <typename Builder>
-cycle_group<Builder> cycle_group<Builder>::unconditional_add(
-    const cycle_group& other, [[maybe_unused]] const std::optional<AffineElement> /*unused*/) const
-    requires IsNotUltraArithmetic<Builder>
-{
-    auto x_diff = other.x - x;
-    auto y_diff = other.y - y;
-    // unconditional add so do not check divisor is zero
-    // (this also makes it much easier to test failure cases as this does not segfault!)
-    auto lambda = y_diff.divide_no_zero_check(x_diff);
-    auto x3 = lambda.madd(lambda, -other.x - x);
-    auto y3 = lambda.madd(x - x3, -y);
-    cycle_group result(x3, y3, /*is_infinity=*/false);
     return result;
 }
 
@@ -838,573 +794,6 @@ template <typename Builder> cycle_group<Builder>& cycle_group<Builder>::operator
     return *this;
 }
 
-template <typename Builder>
-cycle_group<Builder>::cycle_scalar::cycle_scalar(const field_t& _lo, const field_t& _hi)
-    : lo(_lo)
-    , hi(_hi)
-{}
-
-template <typename Builder> cycle_group<Builder>::cycle_scalar::cycle_scalar(const field_t& in)
-{
-    const uint256_t value(in.get_value());
-    const uint256_t lo_v = value.slice(0, LO_BITS);
-    const uint256_t hi_v = value.slice(LO_BITS, HI_BITS);
-    constexpr uint256_t shift = uint256_t(1) << LO_BITS;
-    if (in.is_constant()) {
-        lo = lo_v;
-        hi = hi_v;
-    } else {
-        lo = witness_t(in.get_context(), lo_v);
-        hi = witness_t(in.get_context(), hi_v);
-        (lo + hi * shift).assert_equal(in);
-        // TODO(https://github.com/AztecProtocol/barretenberg/issues/1022): ensure lo and hi are in bb::fr modulus not
-        // bb::fq modulus otherwise we could have two representations for in
-        validate_scalar_is_in_field();
-    }
-    // We need to manually propagate the origin tag
-    lo.set_origin_tag(in.get_origin_tag());
-    hi.set_origin_tag(in.get_origin_tag());
-}
-
-template <typename Builder> cycle_group<Builder>::cycle_scalar::cycle_scalar(const ScalarField& in)
-{
-    const uint256_t value(in);
-    const uint256_t lo_v = value.slice(0, LO_BITS);
-    const uint256_t hi_v = value.slice(LO_BITS, HI_BITS);
-    lo = lo_v;
-    hi = hi_v;
-}
-
-template <typename Builder>
-typename cycle_group<Builder>::cycle_scalar cycle_group<Builder>::cycle_scalar::from_witness(Builder* context,
-                                                                                             const ScalarField& value)
-{
-    const uint256_t value_u256(value);
-    const uint256_t lo_v = value_u256.slice(0, LO_BITS);
-    const uint256_t hi_v = value_u256.slice(LO_BITS, HI_BITS);
-    field_t lo = witness_t(context, lo_v);
-    field_t hi = witness_t(context, hi_v);
-    lo.set_free_witness_tag();
-    hi.set_free_witness_tag();
-    return cycle_scalar(lo, hi);
-}
-
-/**
- * @brief Use when we want to multiply a group element by a string of bits of known size.
- *        N.B. using this constructor method will make our scalar multiplication methods not perform primality tests.
- *
- * @tparam Builder
- * @param context
- * @param value
- * @param num_bits
- * @return cycle_group<Builder>::cycle_scalar
- */
-template <typename Builder>
-typename cycle_group<Builder>::cycle_scalar cycle_group<Builder>::cycle_scalar::from_witness_bitstring(
-    Builder* context, const uint256_t& bitstring, const size_t num_bits)
-{
-    BB_ASSERT_LT(bitstring.get_msb(), num_bits);
-    const uint256_t lo_v = bitstring.slice(0, LO_BITS);
-    const uint256_t hi_v = bitstring.slice(LO_BITS, HI_BITS);
-    field_t lo = witness_t(context, lo_v);
-    field_t hi = witness_t(context, hi_v);
-    lo.set_free_witness_tag();
-    hi.set_free_witness_tag();
-    cycle_scalar result{ lo, hi, num_bits, true, false };
-    return result;
-}
-
-/**
- * @brief Use when we want to multiply a group element by a string of bits of known size.
- *        N.B. using this constructor method will make our scalar multiplication methods not perform primality tests.
- *
- * @tparam Builder
- * @param context
- * @param value
- * @param num_bits
- * @return cycle_group<Builder>::cycle_scalar
- */
-template <typename Builder>
-typename cycle_group<Builder>::cycle_scalar cycle_group<Builder>::cycle_scalar::create_from_bn254_scalar(
-    const field_t& in, const bool skip_primality_test)
-{
-    const uint256_t value_u256(in.get_value());
-    const uint256_t lo_v = value_u256.slice(0, LO_BITS);
-    const uint256_t hi_v = value_u256.slice(LO_BITS, HI_BITS);
-    if (in.is_constant()) {
-        cycle_scalar result{ field_t(lo_v), field_t(hi_v), NUM_BITS, false, true };
-        return result;
-    }
-    field_t lo = witness_t(in.get_context(), lo_v);
-    field_t hi = witness_t(in.get_context(), hi_v);
-    lo.add_two(hi * (uint256_t(1) << LO_BITS), -in).assert_equal(0);
-
-    // We need to manually propagate the origin tag
-    lo.set_origin_tag(in.get_origin_tag());
-    hi.set_origin_tag(in.get_origin_tag());
-
-    cycle_scalar result{ lo, hi, NUM_BITS, skip_primality_test, true };
-    return result;
-}
-/**
- * @brief Construct a new cycle scalar from a bigfield _value, over the same ScalarField Field. If  _value is a witness,
- * we add constraints to ensure the conversion is correct by reconstructing a bigfield from the limbs of the
- * cycle_scalar and checking equality with the initial _value.
- *
- * @tparam Builder
- * @param _value
- * @todo (https://github.com/AztecProtocol/barretenberg/issues/1016): Optimize this method
- */
-template <typename Builder> cycle_group<Builder>::cycle_scalar::cycle_scalar(BigScalarField& scalar)
-{
-    auto* ctx = get_context() ? get_context() : scalar.get_context();
-
-    if (scalar.is_constant()) {
-        const uint256_t value((scalar.get_value() % uint512_t(ScalarField::modulus)).lo);
-        const uint256_t value_lo = value.slice(0, LO_BITS);
-        const uint256_t value_hi = value.slice(LO_BITS, HI_BITS);
-
-        lo = value_lo;
-        hi = value_hi;
-        // N.B. to be able to call assert equal, these cannot be constants
-    } else {
-        // To efficiently convert a bigfield into a cycle scalar,
-        // we are going to explicitly rely on the fact that `scalar.lo` and `scalar.hi`
-        // are implicitly range-constrained to be 128 bits when they are converted into 4-bit lookup window slices
-
-        // First check: can the scalar actually fit into LO_BITS + HI_BITS?
-        // If it can, we can tolerate the scalar being > ScalarField::modulus, because performing a scalar mul
-        // implicilty performs a modular reduction
-        // If not, call `self_reduce` to cut enougn modulus multiples until the above condition is met
-        if (scalar.get_maximum_value() >= (uint512_t(1) << (LO_BITS + HI_BITS))) {
-            scalar.self_reduce();
-        }
-
-        field_t limb0 = scalar.binary_basis_limbs[0].element;
-        field_t limb1 = scalar.binary_basis_limbs[1].element;
-        field_t limb2 = scalar.binary_basis_limbs[2].element;
-        field_t limb3 = scalar.binary_basis_limbs[3].element;
-
-        // The general plan is as follows:
-        // 1. ensure limb0 contains no more than BigScalarField::NUM_LIMB_BITS
-        // 2. define limb1_lo = limb1.slice(0, LO_BITS - BigScalarField::NUM_LIMB_BITS)
-        // 3. define limb1_hi = limb1.slice(LO_BITS - BigScalarField::NUM_LIMB_BITS, <whatever maximum bound of limb1
-        // is>)
-        // 4. construct *this.lo out of limb0 and limb1_lo
-        // 5. construct *this.hi out of limb1_hi, limb2 and limb3
-        // This is a lot of logic, but very cheap on constraints.
-        // For fresh bignums that have come out of a MUL operation,
-        // the only "expensive" part is a size (LO_BITS - BigScalarField::NUM_LIMB_BITS) range check
-
-        // to convert into a cycle_scalar, we need to convert 4*68 bit limbs into 2*128 bit limbs
-        // we also need to ensure that the number of bits in cycle_scalar is < LO_BITS + HI_BITS
-        // note: we do not need to validate that the scalar is within the field modulus
-        // because performing a scalar multiplication implicitly performs a modular reduction (ecc group is
-        // multiplicative modulo BigField::modulus)
-
-        uint256_t limb1_max = scalar.binary_basis_limbs[1].maximum_value;
-
-        // Ensure that limb0 only contains at most NUM_LIMB_BITS. If it exceeds this value, slice of the excess and add
-        // it into limb1
-        if (scalar.binary_basis_limbs[0].maximum_value > BigScalarField::DEFAULT_MAXIMUM_LIMB) {
-            const uint256_t limb = limb0.get_value();
-            const uint256_t lo_v = limb.slice(0, BigScalarField::NUM_LIMB_BITS);
-            const uint256_t hi_v = limb >> BigScalarField::NUM_LIMB_BITS;
-            field_t lo = field_t::from_witness(ctx, lo_v);
-            field_t hi = field_t::from_witness(ctx, hi_v);
-
-            uint256_t hi_max = (scalar.binary_basis_limbs[0].maximum_value >> BigScalarField::NUM_LIMB_BITS);
-            const uint64_t hi_bits = hi_max.get_msb() + 1;
-            lo.create_range_constraint(BigScalarField::NUM_LIMB_BITS);
-            hi.create_range_constraint(static_cast<size_t>(hi_bits));
-            limb0.assert_equal(lo + hi * BigScalarField::shift_1);
-
-            limb1 += hi;
-            limb1_max += hi_max;
-            limb0 = lo;
-        }
-
-        // sanity check that limb[1] is the limb that contributs both to *this.lo and *this.hi
-        BB_ASSERT_GT(BigScalarField::NUM_LIMB_BITS * 2, LO_BITS);
-        BB_ASSERT_LT(BigScalarField::NUM_LIMB_BITS, LO_BITS);
-
-        // limb1 is the tricky one as it contributs to both *this.lo and *this.hi
-        // By this point, we know that limb1 fits in the range `1 << BigScalarField::NUM_LIMB_BITS to  (1 <<
-        // BigScalarField::NUM_LIMB_BITS) + limb1_max.get_maximum_value() we need to slice this limb into 2. The first
-        // is LO_BITS - BigScalarField::NUM_LIMB_BITS (which reprsents its contribution to *this.lo) and the second
-        // represents the limbs contribution to *this.hi Step 1: compute the max bit sizes of both slices
-        const size_t lo_bits_in_limb_1 = LO_BITS - BigScalarField::NUM_LIMB_BITS;
-        const size_t hi_bits_in_limb_1 = (static_cast<size_t>(limb1_max.get_msb()) + 1) - lo_bits_in_limb_1;
-
-        // Step 2: compute the witness values of both slices
-        const uint256_t limb_1 = limb1.get_value();
-        const uint256_t limb_1_hi_multiplicand = (uint256_t(1) << lo_bits_in_limb_1);
-        const uint256_t limb_1_hi_v = limb_1 >> lo_bits_in_limb_1;
-        const uint256_t limb_1_lo_v = limb_1 - (limb_1_hi_v << lo_bits_in_limb_1);
-
-        // Step 3: instantiate both slices as witnesses and validate their sum equals limb1
-        field_t limb_1_lo = field_t::from_witness(ctx, limb_1_lo_v);
-        field_t limb_1_hi = field_t::from_witness(ctx, limb_1_hi_v);
-
-        // We need to propagate the origin tag to the chunks of limb1
-        limb_1_lo.set_origin_tag(limb1.get_origin_tag());
-        limb_1_hi.set_origin_tag(limb1.get_origin_tag());
-        limb1.assert_equal(limb_1_hi * limb_1_hi_multiplicand + limb_1_lo);
-
-        // Step 4: apply range constraints to validate both slices represent the expected contributions to *this.lo and
-        // *this,hi
-        limb_1_lo.create_range_constraint(lo_bits_in_limb_1);
-        limb_1_hi.create_range_constraint(hi_bits_in_limb_1);
-
-        // construct *this.lo out of:
-        // a. `limb0` (the first NUM_LIMB_BITS bits of scalar)
-        // b. `limb_1_lo` (the first LO_BITS - NUM_LIMB_BITS) of limb1
-        lo = limb0 + (limb_1_lo * BigScalarField::shift_1);
-
-        const uint256_t limb_2_shift = uint256_t(1) << (BigScalarField::NUM_LIMB_BITS - lo_bits_in_limb_1);
-        const uint256_t limb_3_shift =
-            uint256_t(1) << ((BigScalarField::NUM_LIMB_BITS - lo_bits_in_limb_1) + BigScalarField::NUM_LIMB_BITS);
-
-        // construct *this.hi out of limb2, limb3 and the remaining term from limb1 not contributing to `lo`
-        hi = limb_1_hi.add_two(limb2 * limb_2_shift, limb3 * limb_3_shift);
-    }
-    // We need to manually propagate the origin tag
-    lo.set_origin_tag(scalar.get_origin_tag());
-    hi.set_origin_tag(scalar.get_origin_tag());
-};
-
-template <typename Builder> bool cycle_group<Builder>::cycle_scalar::is_constant() const
-{
-    return (lo.is_constant() && hi.is_constant());
-}
-
-/**
- * @brief Checks that a cycle_scalar value is smaller than a prime field modulus when evaluated over the INTEGERS
- * N.B. The prime we check can be either the SNARK curve group order or the circuit's embedded curve group order
- * (i.e. BN254 or Grumpkin)
- * For a canonical scalar mul, we check against the embedded curve (i.e. the curve
- * cycle_group implements).
- * HOWEVER: for Pedersen hashes and Pedersen commitments, the hashed/committed data will be
- * native circuit field elements i.e. for a BN254 snark, cycle_group = Grumpkin and we will be committing/hashing
- * BN254::ScalarField values *NOT* Grumpkin::ScalarFIeld values.
- * TLDR: whether the input scalar has to be < BN254::ScalarField or < Grumpkin::ScalarField is context-dependent.
- *
- * @tparam Builder
- */
-template <typename Builder> void cycle_group<Builder>::cycle_scalar::validate_scalar_is_in_field() const
-{
-    if (!is_constant() && !skip_primality_test()) {
-        // Check that scalar.hi * 2^LO_BITS + scalar.lo < cycle_group_modulus when evaluated over the integers
-        const uint256_t cycle_group_modulus =
-            use_bn254_scalar_field_for_primality_test() ? FF::modulus : ScalarField::modulus;
-        const uint256_t r_lo = cycle_group_modulus.slice(0, cycle_scalar::LO_BITS);
-        const uint256_t r_hi = cycle_group_modulus.slice(cycle_scalar::LO_BITS, cycle_scalar::HI_BITS);
-
-        bool need_borrow = uint256_t(lo.get_value()) > r_lo;
-        field_t borrow = lo.is_constant() ? need_borrow : field_t::from_witness(get_context(), need_borrow);
-
-        // directly call `create_new_range_constraint` to avoid creating an arithmetic gate
-        if (!lo.is_constant()) {
-            // We need to manually propagate the origin tag
-            borrow.set_origin_tag(lo.get_origin_tag());
-            if constexpr (IS_ULTRA) {
-                get_context()->create_new_range_constraint(borrow.get_witness_index(), 1, "borrow");
-            } else {
-                borrow.assert_equal(borrow * borrow);
-            }
-        }
-        // Hi range check = r_hi - y_hi - borrow
-        // Lo range check = r_lo - y_lo + borrow * 2^{126}
-        field_t hi_diff = (-hi + r_hi) - borrow;
-        field_t lo_diff = (-lo + r_lo) + (borrow * (uint256_t(1) << cycle_scalar::LO_BITS));
-
-        hi_diff.create_range_constraint(cycle_scalar::HI_BITS);
-        lo_diff.create_range_constraint(cycle_scalar::LO_BITS);
-    }
-}
-
-template <typename Builder>
-typename cycle_group<Builder>::ScalarField cycle_group<Builder>::cycle_scalar::get_value() const
-{
-    uint256_t lo_v(lo.get_value());
-    uint256_t hi_v(hi.get_value());
-    return ScalarField(lo_v + (hi_v << LO_BITS));
-}
-
-/**
- * @brief Construct a new cycle group<Builder>::straus scalar slice::straus scalar slice object
- *
- * @details As part of slicing algoirthm, we also perform a primality test on the inut scalar.
- *
- * TODO(@zac-williamson) make the primality test configurable.
- * We may want to validate the input < BN254::Fr OR input < Grumpkin::Fr depending on context!
- *
- * @tparam Builder
- * @param context
- * @param scalar
- * @param table_bits
- */
-template <typename Builder>
-cycle_group<Builder>::straus_scalar_slice::straus_scalar_slice(Builder* context,
-                                                               const cycle_scalar& scalar,
-                                                               const size_t table_bits)
-    : _table_bits(table_bits)
-{
-    // convert an input cycle_scalar object into a vector of slices, each containing `table_bits` bits.
-    // this also performs an implicit range check on the input slices
-    const auto slice_scalar = [&](const field_t& scalar, const size_t num_bits) {
-        // we record the scalar slices both as field_t circuit elements and u64 values
-        // (u64 values are used to index arrays and we don't want to repeatedly cast a stdlib value to a numeric
-        // primitive as this gets expensive when repeated enough times)
-        std::pair<std::vector<field_t>, std::vector<uint64_t>> result;
-        result.first.reserve(static_cast<size_t>(1ULL) << table_bits);
-        result.second.reserve(static_cast<size_t>(1ULL) << table_bits);
-
-        if (num_bits == 0) {
-            return result;
-        }
-        if (scalar.is_constant()) {
-            const size_t num_slices = (num_bits + table_bits - 1) / table_bits;
-            const uint64_t table_mask = (1ULL << table_bits) - 1ULL;
-            uint256_t raw_value = scalar.get_value();
-            for (size_t i = 0; i < num_slices; ++i) {
-                uint64_t slice_v = static_cast<uint64_t>(raw_value.data[0]) & table_mask;
-                result.first.push_back(field_t(slice_v));
-                result.second.push_back(slice_v);
-                raw_value = raw_value >> table_bits;
-            }
-
-            return result;
-        }
-        uint256_t raw_value = scalar.get_value();
-        const uint64_t table_mask = (1ULL << table_bits) - 1ULL;
-        const size_t num_slices = (num_bits + table_bits - 1) / table_bits;
-        for (size_t i = 0; i < num_slices; ++i) {
-            uint64_t slice_v = static_cast<uint64_t>(raw_value.data[0]) & table_mask;
-            result.second.push_back(slice_v);
-            raw_value = raw_value >> table_bits;
-        }
-
-        if constexpr (IS_ULTRA) {
-            const auto slice_indices =
-                context->decompose_into_default_range(scalar.get_normalized_witness_index(),
-                                                      num_bits,
-                                                      table_bits,
-                                                      "straus_scalar_slice decompose_into_default_range");
-            for (auto& idx : slice_indices) {
-                result.first.emplace_back(field_t::from_witness_index(context, idx));
-            }
-        } else {
-            for (size_t i = 0; i < num_slices; ++i) {
-                uint64_t slice_v = result.second[i];
-                field_t slice(witness_t(context, slice_v));
-
-                context->create_range_constraint(
-                    slice.get_witness_index(), table_bits, "straus_scalar_slice create_range_constraint");
-
-                result.first.push_back(slice);
-            }
-            std::vector<field_t> linear_elements;
-            FF scaling_factor = 1;
-            for (size_t i = 0; i < num_slices; ++i) {
-                linear_elements.emplace_back(result.first[i] * scaling_factor);
-                scaling_factor += scaling_factor;
-            }
-            field_t::accumulate(linear_elements).assert_equal(scalar);
-        }
-        return result;
-    };
-
-    const size_t lo_bits = scalar.num_bits() > cycle_scalar::LO_BITS ? cycle_scalar::LO_BITS : scalar.num_bits();
-    const size_t hi_bits = scalar.num_bits() > cycle_scalar::LO_BITS ? scalar.num_bits() - cycle_scalar::LO_BITS : 0;
-    auto hi_slices = slice_scalar(scalar.hi, hi_bits);
-    auto lo_slices = slice_scalar(scalar.lo, lo_bits);
-
-    std::copy(lo_slices.first.begin(), lo_slices.first.end(), std::back_inserter(slices));
-    std::copy(hi_slices.first.begin(), hi_slices.first.end(), std::back_inserter(slices));
-    std::copy(lo_slices.second.begin(), lo_slices.second.end(), std::back_inserter(slices_native));
-    std::copy(hi_slices.second.begin(), hi_slices.second.end(), std::back_inserter(slices_native));
-    const auto tag = scalar.get_origin_tag();
-    for (auto& element : slices) {
-        // All slices need to have the same origin tag
-        element.set_origin_tag(tag);
-    }
-}
-
-/**
- * @brief Return a bit-slice associated with round `index`.
- *
- * @details In Straus algorithm, `index` is a known parameter, so no need for expensive lookup tables
- *
- * @tparam Builder
- * @param index
- * @return field_t<Builder>
- */
-template <typename Builder>
-std::optional<field_t<Builder>> cycle_group<Builder>::straus_scalar_slice::read(size_t index)
-{
-    if (index >= slices.size()) {
-        return std::nullopt;
-    }
-    return slices[index];
-}
-
-/**
- * @brief Compute the output points generated when computing the Straus lookup table
- * @details When performing an MSM, we first compute all the witness values as Element types (with a Z-coordinate),
- *          and then we batch-convert the points into affine representation `AffineElement`
- *          This avoids the need to compute a modular inversion for every group operation,
- *          which dramatically cuts witness generation times
- *
- * @tparam Builder
- * @param base_point
- * @param offset_generator
- * @param table_bits
- * @return std::vector<typename cycle_group<Builder>::Element>
- */
-template <typename Builder>
-std::vector<typename cycle_group<Builder>::Element> cycle_group<
-    Builder>::straus_lookup_table::compute_straus_lookup_table_hints(const Element& base_point,
-                                                                     const Element& offset_generator,
-                                                                     size_t table_bits)
-{
-    const size_t table_size = 1UL << table_bits;
-    Element base = base_point.is_point_at_infinity() ? Group::one : base_point;
-    std::vector<Element> hints;
-    hints.emplace_back(offset_generator);
-    for (size_t i = 1; i < table_size; ++i) {
-        hints.emplace_back(hints[i - 1] + base);
-    }
-    return hints;
-}
-
-/**
- * @brief Construct a new cycle group<Builder>::straus lookup table::straus lookup table object
- *
- * @details Constructs a `table_bits` lookup table.
- *
- * If Builder is not ULTRA, `table_bits = 1`
- * If Builder is ULTRA, ROM table is used as lookup table
- *
- * @tparam Builder
- * @param context
- * @param base_point
- * @param offset_generator
- * @param table_bits
- */
-template <typename Builder>
-cycle_group<Builder>::straus_lookup_table::straus_lookup_table(Builder* context,
-                                                               const cycle_group& base_point,
-                                                               const cycle_group& offset_generator,
-                                                               size_t table_bits,
-                                                               std::optional<std::span<AffineElement>> hints)
-    : _table_bits(table_bits)
-    , _context(context)
-    , tag(OriginTag(base_point.get_origin_tag(), offset_generator.get_origin_tag()))
-{
-    const size_t table_size = 1UL << table_bits;
-    point_table.resize(table_size);
-    point_table[0] = offset_generator;
-
-    // We want to support the case where input points are points at infinity.
-    // If base point is at infinity, we want every point in the table to just be `generator_point`.
-    // We achieve this via the following:
-    // 1: We create a "work_point" that is base_point if not at infinity, otherwise is just 1
-    // 2: When computing the point table, we use "work_point" in additions instead of the "base_point" (to prevent
-    //    x-coordinate collisions in honest case) 3: When assigning to the point table, we conditionally assign either
-    //    the output of the point addition (if not at infinity) or the generator point (if at infinity)
-    // Note: if `base_point.is_point_at_infinity()` is constant, these conditional assigns produce zero gate overhead
-    cycle_group fallback_point(Group::affine_one);
-    field_t modded_x = field_t::conditional_assign(base_point.is_point_at_infinity(), fallback_point.x, base_point.x);
-    field_t modded_y = field_t::conditional_assign(base_point.is_point_at_infinity(), fallback_point.y, base_point.y);
-    cycle_group modded_base_point(modded_x, modded_y, false);
-
-    // if the input point is constant, it is cheaper to fix the point as a witness and then derive the table, than it is
-    // to derive the table and fix its witnesses to be constant! (due to group additions = 1 gate, and fixing x/y coords
-    // to be constant = 2 gates)
-    if (base_point.is_constant() && !base_point.is_point_at_infinity().get_value()) {
-        modded_base_point = cycle_group::from_constant_witness(_context, modded_base_point.get_value());
-        point_table[0] = cycle_group::from_constant_witness(_context, offset_generator.get_value());
-        for (size_t i = 1; i < table_size; ++i) {
-            std::optional<AffineElement> hint =
-                hints.has_value() ? std::optional<AffineElement>(hints.value()[i - 1]) : std::nullopt;
-            point_table[i] = point_table[i - 1].unconditional_add(modded_base_point, hint);
-        }
-    } else {
-        std::vector<std::tuple<field_t, field_t>> x_coordinate_checks;
-        // ensure all of the ecc add gates are lined up so that we can pay 1 gate per add and not 2
-        for (size_t i = 1; i < table_size; ++i) {
-            std::optional<AffineElement> hint =
-                hints.has_value() ? std::optional<AffineElement>(hints.value()[i - 1]) : std::nullopt;
-            x_coordinate_checks.emplace_back(point_table[i - 1].x, modded_base_point.x);
-            point_table[i] = point_table[i - 1].unconditional_add(modded_base_point, hint);
-        }
-
-        // batch the x-coordinate checks together
-        // because `assert_is_not_zero` witness generation needs a modular inversion (expensive)
-        field_t coordinate_check_product = 1;
-        for (auto& [x1, x2] : x_coordinate_checks) {
-            auto x_diff = x2 - x1;
-            coordinate_check_product *= x_diff;
-        }
-        coordinate_check_product.assert_is_not_zero("straus_lookup_table x-coordinate collision");
-
-        for (size_t i = 1; i < table_size; ++i) {
-            point_table[i] =
-                cycle_group::conditional_assign(base_point.is_point_at_infinity(), offset_generator, point_table[i]);
-        }
-    }
-    if constexpr (IS_ULTRA) {
-        rom_id = context->create_ROM_array(table_size);
-        for (size_t i = 0; i < table_size; ++i) {
-            if (point_table[i].is_constant()) {
-                auto element = point_table[i].get_value();
-                point_table[i] = cycle_group::from_constant_witness(_context, element);
-                point_table[i].x.assert_equal(element.x);
-                point_table[i].y.assert_equal(element.y);
-            }
-            context->set_ROM_element_pair(
-                rom_id,
-                i,
-                std::array<uint32_t, 2>{ point_table[i].x.get_witness_index(), point_table[i].y.get_witness_index() });
-        }
-    } else {
-        BB_ASSERT_EQ(table_bits, 1U);
-    }
-}
-
-/**
- * @brief Given an `_index` witness, return `straus_lookup_table[index]`
- *
- * @tparam Builder
- * @param _index
- * @return cycle_group<Builder>
- */
-template <typename Builder> cycle_group<Builder> cycle_group<Builder>::straus_lookup_table::read(const field_t& _index)
-{
-    if constexpr (IS_ULTRA) {
-        field_t index(_index);
-        if (index.is_constant()) {
-            index = witness_t(_context, _index.get_value());
-            index.assert_equal(_index.get_value());
-        }
-        auto output_indices = _context->read_ROM_array_pair(rom_id, index.get_witness_index());
-        field_t x = field_t::from_witness_index(_context, output_indices[0]);
-        field_t y = field_t::from_witness_index(_context, output_indices[1]);
-        // Merge tag of table with tag of index
-        x.set_origin_tag(OriginTag(tag, _index.get_origin_tag()));
-        y.set_origin_tag(OriginTag(tag, _index.get_origin_tag()));
-        return cycle_group(x, y, /*is_infinity=*/false);
-    }
-    field_t x = _index * (point_table[1].x - point_table[0].x) + point_table[0].x;
-    field_t y = _index * (point_table[1].y - point_table[0].y) + point_table[0].y;
-
-    // Merge tag of table with tag of index
-    x.set_origin_tag(OriginTag(tag, _index.get_origin_tag()));
-    y.set_origin_tag(OriginTag(tag, _index.get_origin_tag()));
-    return cycle_group(x, y, /*is_infinity=*/false);
-}
-
 /**
  * @brief Internal algorithm to perform a variable-base batch mul.
  *
@@ -1696,94 +1085,6 @@ typename cycle_group<Builder>::batch_mul_internal_output cycle_group<Builder>::_
      */
     // Set accumulator's origin tag to the union of all scalars' tags
     accumulator.set_origin_tag(tag);
-    return { accumulator, offset_generator_accumulator };
-}
-
-/**
- * @brief Internal algorithm to perform a fixed-base batch mul for Non-ULTRA Builders
- *
- * @details Multiples of the base point are precomputed, which avoids us having to add ecc doubling gates.
- *          More efficient than variable-base version.
- *
- * @tparam Builder
- * @param scalars
- * @param base_points
- * @param off
- * @return cycle_group<Builder>::batch_mul_internal_output
- */
-template <typename Builder>
-typename cycle_group<Builder>::batch_mul_internal_output cycle_group<Builder>::_fixed_base_batch_mul_internal(
-    const std::span<cycle_scalar> scalars,
-    const std::span<AffineElement> base_points,
-    const std::span<AffineElement const> offset_generators)
-    requires IsNotUltraArithmetic<Builder>
-
-{
-    BB_ASSERT_EQ(scalars.size(), base_points.size());
-    static_assert(TABLE_BITS == 1);
-
-    Builder* context = nullptr;
-    for (auto& scalar : scalars) {
-        if (scalar.get_context() != nullptr) {
-            context = scalar.get_context();
-            break;
-        }
-    }
-
-    size_t num_bits = 0;
-    for (auto& s : scalars) {
-        num_bits = std::max(num_bits, s.num_bits());
-    }
-    size_t num_rounds = (num_bits + TABLE_BITS - 1) / TABLE_BITS;
-    // core algorithm
-    // define a `table_bits` size lookup table
-    const size_t num_points = scalars.size();
-    using straus_round_tables = std::vector<straus_lookup_table>;
-
-    std::vector<straus_scalar_slice> scalar_slices;
-    std::vector<straus_round_tables> point_tables(num_points);
-
-    // creating these point tables should cost 0 constraints if base points are constant
-    for (size_t i = 0; i < num_points; ++i) {
-        std::vector<Element> round_points(num_rounds);
-        std::vector<Element> round_offset_generators(num_rounds);
-        round_points[0] = base_points[i];
-        round_offset_generators[0] = offset_generators[i + 1];
-        for (size_t j = 1; j < num_rounds; ++j) {
-            round_points[j] = round_points[j - 1].dbl();
-            round_offset_generators[j] = round_offset_generators[j - 1].dbl();
-        }
-        Element::batch_normalize(&round_points[0], num_rounds);
-        Element::batch_normalize(&round_offset_generators[0], num_rounds);
-        point_tables[i].resize(num_rounds);
-        for (size_t j = 0; j < num_rounds; ++j) {
-            point_tables[i][j] = straus_lookup_table(
-                context, cycle_group(round_points[j]), cycle_group(round_offset_generators[j]), TABLE_BITS);
-        }
-        scalar_slices.emplace_back(straus_scalar_slice(context, scalars[i], TABLE_BITS));
-    }
-    Element offset_generator_accumulator = offset_generators[0];
-    cycle_group accumulator = cycle_group(Element(offset_generators[0]) * (uint256_t(1) << (num_rounds - 1)));
-    for (size_t i = 0; i < num_rounds; ++i) {
-        offset_generator_accumulator = (i > 0) ? offset_generator_accumulator.dbl() : offset_generator_accumulator;
-        for (size_t j = 0; j < num_points; ++j) {
-            auto& point_table = point_tables[j][i];
-            const std::optional<field_t> scalar_slice = scalar_slices[j].read(i);
-            // if we are doing a batch mul over scalars of different bit-lengths, we may not have any scalar bits for a
-            // given round and a given scalar
-            if (scalar_slice.has_value()) {
-                const cycle_group point = point_table.read(scalar_slice.value());
-                accumulator = accumulator.unconditional_add(point);
-                offset_generator_accumulator = offset_generator_accumulator + Element(offset_generators[j + 1]);
-            }
-        }
-    }
-
-    /**
-     * offset_generator_accumulator represents the sum of all the offset generator terms present in `accumulator`.
-     * We don't subtract off yet, as we may be able to combine `offset_generator_accumulator` with other constant terms
-     * in `batch_mul` before performing the subtraction.
-     */
     return { accumulator, offset_generator_accumulator };
 }
